requirements.txt
# OpenAI API
openai>=1.0.0,<3.0.0

# Environment variable management
python-dotenv>=1.0.0,<1.1.0

# Data handling
pandas>=2.1.0,<3.0.0
numpy>=1.25.0,<2.0.0
python-dateutil>=2.8.2,<3.0.0
pytz>=2023.3,<2024.0

# Visualization
matplotlib>=3.7.2,<4.0.0
plotly>=6.0.0,<7.0.0
seaborn>=0.12.2,<0.13.0
Pillow>=10.1.0,<11.0.0

# Web interface
gradio>=3.46.0,<4.0.0

# Logging and utilities
loguru>=0.7.3,<0.8.0
rich>=13.0.0,<14.0.0

# Optional: HTTP requests (if any agent uses it)
requests>=2.32.0,<3.0.0
protobuf<5.0.0  # safer for compatibility with Gradio
